fix(build): [STABLE] npm publish readiness — cache dir, build config, stale expected.c
- CACHE_DIR: move from ROOT/.tsclang-cache to ~/.tsclang/cache (writable on global install)
  - Support TSCLANG_CACHE_DIR env var override
  - Remove unused ROOT/fileURLToPath imports from compile.ts
- tsconfig.build.json: production build (no sourcemap, no declaration)
- package.json: build script uses tsconfig.build.json
- .gitignore: remove .tsclang-cache/ (no longer in project root)
- Regenerate 9 stale expected.c files (safe-math uint32 cast was masked by old cache)
- 1764 tests pass, global install verified (build + run)

+++ test/cases/14-stdlib/reactive-library/pkg-import/expected.c
@@ -1,11 +1,11 @@
 #include "runtime.h"
 
 int32_t mymath_add_i32_i32(int32_t a, int32_t b) {
-    return a + b;
+    return (int32_t)((uint32_t)a + (uint32_t)b);
 }
 
 int32_t mymath_mul_i32_i32(int32_t a, int32_t b) {
-    return a * b;
+    return (int32_t)((uint32_t)a * (uint32_t)b);
 }
 
 int main(void) {
